refactor: move routes to new file
